changed the structure to makes tests separate project
