+++ CodeGen/examples/Package.xml
@@ -1,0 +1,96 @@
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<Package>
+    <name>main</name>
+    <classes>
+        <class>
+            <name>Fruit</name>
+        </class>
+        <class>
+            <name>Apple</name>
+            <fields>
+                <field>
+                    <name>colour</name>
+                    <type>string</type>
+                    <const>false</const>
+                    <pointer>false</pointer>
+                    <default>&#34;red&#34;</default>
+                    <access>public</access>
+                    <static>false</static>
+                </field>
+                <field>
+                    <name>sort</name>
+                    <type>string</type>
+                    <const>false</const>
+                    <pointer>false</pointer>
+                    <default>&#34;Golden&#34;</default>
+                    <access>public</access>
+                    <static>true</static>
+                </field>
+                <field>
+                    <name>size</name>
+                    <type>int</type>
+                    <const>false</const>
+                    <pointer>false</pointer>
+                    <default>1</default>
+                    <access>private</access>
+                    <static>false</static>
+                </field>
+            </fields>
+            <methods>
+                <method>
+                    <name>print</name>
+                    <access>private</access>
+                    <const>false</const>
+                    <static>false</static>
+                    <parameters>
+                        <parameter>
+                            <name>colour</name>
+                            <type>string</type>
+                        </parameter>
+                    </parameters>
+                </method>
+                <method>
+                    <name>getSize</name>
+                    <return>int</return>
+                    <access>protected</access>
+                    <const>false</const>
+                    <static>true</static>
+                </method>
+                <method>
+                    <name>getColor</name>
+                    <return>string</return>
+                    <access>public</access>
+                    <const>true</const>
+                    <static>false</static>
+                </method>
+            </methods>
+            <classes>
+                <class>
+                    <name>Seed</name>
+                    <fields>
+                        <field>
+                            <name>size</name>
+                            <type>int</type>
+                            <const>false</const>
+                            <pointer>false</pointer>
+                            <access>public</access>
+                            <static>false</static>
+                        </field>
+                    </fields>
+                    <methods>
+                        <method>
+                            <name>transform</name>
+                            <return>int</return>
+                            <access>public</access>
+                            <const>true</const>
+                            <static>true</static>
+                        </method>
+                    </methods>
+                    <access>private</access>
+                </class>
+            </classes>
+            <parent>Fruit</parent>
+        </class>
+    </classes>
+    <use_spaces>false</use_spaces>
+</Package>
